backend/auth_service/requirements.txt
fastapi
uvicorn
motor
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic-settings
